[feature/create-new-wallet] Update: save info when getAccountInfo, done.
