+++ src/navigation/AuthenticationNavigation/AuthenticationRouter.ts
@@ -1,5 +1,6 @@
 const AuthenticationRouter = {
     WELCOME_SCREEN: 'WELCOME_SCREEN',
-    CREATE_NEW_WALLET: 'CREATE_NEW_WALLET'
+    CREATE_NEW_WALLET: 'CREATE_NEW_WALLET',
+    CHOOSE_PIN: 'CHOOSE_PIN',
 };
 export default AuthenticationRouter;
