+++ src/navigation/ChoosePinNavigation/ChoosePinRouter.ts
@@ -1,0 +1,5 @@
+const ChoosePinRouter = {
+    CHOOSE_PIN_SCREEN: 'CHOOSE_PIN_SCREEN',
+    CONFIRM_PIN_SCREEN: 'CONFIRM_PIN_SCREEN',
+};
+export default ChoosePinRouter;
