+++ src/navigation/CreateNewWalletNavigation/CreateNewWalletRouter.ts
@@ -3,10 +3,6 @@ const CreateNewWalletRouter = {
     RECOVERY_PHRASE_SCREEN: 'RECOVERY_PHRASE_SCREEN',
     DOUBLE_CHECK_IT_SCREEN: 'DOUBLE_CHECK_IT_SCREEN',
     IMPORT_PHRASE_SCREEN: 'IMPORT_PHRASE_SCREEN',
-    CHOOSE_PIN_SCREEN: 'CHOOSE_PIN_SCREEN',
-    CONFIRM_PIN_SCREEN: 'CONFIRM_PIN_SCREEN',
-    DEVICE_SELECTION_SCREEN: 'DEVICE_SELECTION_SCREEN',
-    SHOW_ADDRESS_SCREEN: 'SHOW_ADDRESS_SCREEN',
     CONNECT_LEDGER_SCREEN: 'CONNECT_LEDGER_SCREEN',
 };
