+++ src/screens/authentication/create_new_wallet/DoubleCheckItScreen/index.tsx
@@ -3,20 +3,28 @@ import {ScrollView, StyleSheet} from 'react-native';
 import {ScreenProps} from 'navigation/ScreenProps';
 import CreateNewWalletRouter from 'navigation/CreateNewWalletNavigation/CreateNewWalletRouter';
 import _ from 'lodash';
-import {CLayout, CHeader} from 'components';
+import {CLayout, CHeader, CLoading} from 'components';
 import {scale} from 'device';
 import CButton2 from 'components/CTextButton';
 import {useNavigation} from '@react-navigation/native';
 import {CheckItem} from 'screens/authentication/create_new_wallet/components';
 import {StackNavigationProp} from '@react-navigation/stack';
 import {getArrayNotInArray} from 'utils/collections';
+import {allActions} from 'redux_manager';
+import {Config} from 'utils';
+import {useDispatch} from 'react-redux';
+import {Phrase} from 'screens/authentication/data/data';
+import Keys from 'utils/keys';
+import AuthenticationRouter from 'navigation/AuthenticationNavigation/AuthenticationRouter';
 
 // @ts-ignore
 const DoubleCheckItScreen: React.FC<ScreenProps<CreateNewWalletRouter.DOUBLE_CHECK_IT_SCREEN>> = ({route}) => {
     const [listData, setListData] = useState<any>([]);
     const [listDataSelected, setListDataSelected] = useState<any>([]);
+    const [isLoading, setLoading] = useState<boolean>(false);
     const numberOfWords = Math.floor(route.params.data.length / 3);
-    const {navigate} = useNavigation<StackNavigationProp<any>>();
+    const {replace} = useNavigation<StackNavigationProp<any>>();
+    const dispatch = useDispatch();
 
     useEffect(() => {
         const {data} = route.params;
@@ -56,8 +64,28 @@ const DoubleCheckItScreen: React.FC<ScreenProps<CreateNewWalletRouter.DOUBLE_CHE
         }
     };
 
-    const openChoosePin = () => {
-        navigate(CreateNewWalletRouter.CHOOSE_PIN_SCREEN);
+    const getAccountInformation = () => {
+        const {data} = route.params;
+        const phrases = data.reduce((previous: string, current: Phrase) => previous + current.word + ' ', '');
+        const publicKey = '0160d88b3f847221f4dc6c5549dcfc26772c02f253a24de226a88b4536bc61d4ad'; //TODO: get publicKey from phrases string
+        setLoading(true);
+        dispatch(allActions.user.getAccountInformation(publicKey, async (err: any, res: any) => {
+            if (res) {
+                setLoading(false);
+                const info = {
+                    publicKey: publicKey,
+                    loginOptions: {
+                        connectionType: 'passphase',
+                        passphase: phrases,
+                    },
+                };
+                await Config.saveItem(Keys.casperdash, info);
+                replace(AuthenticationRouter.CHOOSE_PIN);
+            } else {
+                setLoading(false);
+                Config.alertMess(err);
+            }
+        }));
     };
 
     return (
@@ -79,11 +107,12 @@ const DoubleCheckItScreen: React.FC<ScreenProps<CreateNewWalletRouter.DOUBLE_CHE
                 }
             </ScrollView>
             <CButton2
-                onPress={openChoosePin}
+                onPress={getAccountInformation}
                 style={styles.btnNext}
                 disabled={listDataSelected.filter((i: any) => !!i).length !== numberOfWords}
                 text={'Next'}
             />
+            {isLoading && <CLoading/>}
         </CLayout>
     );
 };
