--- src/types/choose.d.ts
@@ -1,6 +1,0 @@
-interface IChoose {
-  id: number;
-  monthly: number;
-  price: number;
-  discount?: number;
-}
