--- src/types/method.d.ts
@@ -1,5 +1,0 @@
-interface IMethod {
-  id: number;
-  name: string;
-  value: string;
-}
