+++ src/utils/index.ts
@@ -1,1 +1,2 @@
-export { default as Config } from './config';
+export {default as Config} from './config';
+export {default as Keys} from './keys';
