+++ src/utils/keys.tsx
@@ -1,0 +1,5 @@
+const Keys = {
+    casperdash: 'casperdash',
+};
+
+export default Keys;
